# openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/student": {
      "post": {
        "operationId": "StudentController_create",
        "summary": "Create student",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been sucessfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentPreviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "Student"
        ]
      },
      "get": {
        "operationId": "StudentController_findAll",
        "summary": "Get students with pagination and search",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Search by name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "required": false,
            "in": "query",
            "description": "Search by description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "university",
            "required": false,
            "in": "query",
            "description": "Include university data in the result",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "professor",
            "required": false,
            "in": "query",
            "description": "Include professor data in the result",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The records has been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentPreviewDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Student"
        ]
      }
    },
    "/student/{id}": {
      "get": {
        "operationId": "StudentController_findOne",
        "summary": "Get student by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record has been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              }
            }
          }
        },
        "tags": [
          "Student"
        ]
      },
      "patch": {
        "operationId": "StudentController_update",
        "summary": "Update student by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The record has been successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              }
            }
          },
          "404": {
            "description": "Student with id not found"
          }
        },
        "tags": [
          "Student"
        ]
      },
      "delete": {
        "operationId": "StudentController_remove",
        "summary": "Delete student by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully deleted"
          }
        },
        "tags": [
          "Student"
        ]
      }
    },
...
    "/university": {
      "post": {
        "operationId": "UniversityController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUniversityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "UniversityController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/university/{id}": {
      "get": {
        "operationId": "UniversityController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "UniversityController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUniversityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "UniversityController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "YAssignment",
    "description": "YAssignment's API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "professor": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          }
        },
        "required": [
          "name",
          "description",
          "university",
          "professor"
        ]
      },
      "StudentPreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          },
          "professor": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
      "StudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          },
          "professor": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "university",
          "professor"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "professor": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          }
        }
      },
...
      "CreateUniversityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "students": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "format": "uuid",
              "minLength": 1
            }
          },
          "professors": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "format": "uuid",
              "minLength": 1
            }
          }
        },
        "required": [
          "name",
          "description",
          "students",
          "professors"
        ]
      },
      "UpdateUniversityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          }
        }
      }
    }
  }
}